Switch to calculate_time_remaining
